[Refactor]: move agent runtime into package
