+++ ARCHITECTURE.md
@@ -26,7 +26,7 @@
   ┌──────────────────────────┐
   │ Agent Runtime            │
   │ ian.services             │
-  │ - agent_runtime          │
+  │ - agent/                 │
   │ - LangGraph ReAct        │
   │ - Gemini 3 Flash         │
   └──────────────────────────┘
@@ -35,7 +35,7 @@
 ### 各層職責
 
 - **Gateway 層**：各平台入口。`ian.gateways.discord_bot` 處理 Discord Slash Commands；`ian.gateways.webhook_server` (Flask) 負責 Webhook route wiring，並委派給 `ian.gateways.facebook_webhook` 與 `ian.gateways.line_webhook` 處理 Facebook Messenger / LINE 平台細節。
-- **Host Agent Client**：`ian.services.agent_runtime` 使用 LangGraph `create_react_agent` 搭配 Google Gemini 3 Flash，透過 MCP 協定調用工具，並管理每位使用者的獨立對話 session。
+- **Host Agent Client**：`ian.services.agent` 使用 LangGraph `create_react_agent` 搭配 Google Gemini 3 Flash，透過 MCP 協定調用工具，並管理每位使用者的獨立對話 session。
 - **MCP Tool Server**：`ian.gateways.mcp_server` 以 FastMCP 框架透過 SSE 提供 RAG 搜尋、課程查詢、幹部通知、社員綁定、簽到碼產生、訂閱管理、個性備註等工具。
 - **Member DB**：`ian.services.member_store` 從 Google Apps Script API 同步社員資料至本地 JSON 快取，提供平台帳號查詢、角色辨識、Email 綁定、訂閱管理與個性備註功能。
 
@@ -72,7 +72,7 @@ ntuai-watson-agent/
 
 ## 核心元件
 
-### Host Agent Client (`ian.services.agent_runtime`)
+### Host Agent Client (`ian.services.agent`)
 
 - 使用 **LangGraph** `create_react_agent` 搭配 **Google Gemini 3 Flash** (`gemini-3-flash-preview`) 建立 ReAct 推理迴圈。
 - 每位使用者擁有獨立 session（含 `MemorySaver` 對話記憶），閒置 15 分鐘自動過期。
