+++ src/ian/gateways/facebook_webhook.py
@@ -11,7 +11,7 @@ from ian.gateways.messaging_common import (
     get_current_time,
     save_chat_history,
 )
-from ian.services.agent_runtime import chat_with_agent, parse_no_response, start_dispatcher
+from ian.services.agent import chat_with_agent, parse_no_response, start_dispatcher
 from ian.services.member_store import (
     get_member_name as get_member_name_from_db,
     get_member_role as get_member_role_from_db,
