+++ src/ian/services/agent/__init__.py
@@ -1,0 +1,23 @@
+"""Agent runtime public API."""
+
+from ian.domain.urls import parse_no_response
+from ian.services.agent.logging import (
+    add_log,
+    send_startup_notification,
+    start_log_processor,
+)
+from ian.services.agent.runtime import (
+    chat_with_agent,
+    start_dispatcher,
+)
+from ian.services.agent.sessions import clear_session
+
+__all__ = [
+    "add_log",
+    "chat_with_agent",
+    "clear_session",
+    "parse_no_response",
+    "send_startup_notification",
+    "start_dispatcher",
+    "start_log_processor",
+]
